requirements.txt
pytesseract
pdf2image
Pillow
google-genai
gradio
pymupdf
pandas
